third_party/WebKit/LayoutTests/inspector/tracing/timeline-style-recalc-invalidations-expected.txt

Tests the Timeline API instrumentation of style recalc invalidations.

PASS


Running: testLocalFrame

Running: testSubframe
